Finance: rename maxRepeats to maxExecutions (#739)
Renames `repeats` to `executions`, `maxRepeats` to `maxExecutions`, and `paymentRepeats` to `paymentExecutions` to better reflect their parameters and states variables' meaning.

In particular, a `maxRepeats` of 1 is quite confusing as to whether it means you can repeat the payment once (e.g. actually execute it twice), or only execute it once.